Pagination divided class
